refactor : 카테고리(목록) 조회 용 Service 분리 (#803)
