+++ src/main/java/com/clubber/ClubberServer/domain/club/controller/ClubController.java
@@ -10,6 +10,7 @@ import com.clubber.ClubberServer.domain.club.dto.GetClubPopularResponse;
 import com.clubber.ClubberServer.domain.club.dto.GetClubResponse;
 import com.clubber.ClubberServer.domain.club.dto.GetOfficialClubGroupResponse;
 import com.clubber.ClubberServer.domain.club.dto.GetSummaryClubGroupResponse;
+import com.clubber.ClubberServer.domain.club.service.ClubCategoyService;
 import com.clubber.ClubberServer.domain.club.service.ClubService;
 import com.clubber.ClubberServer.global.config.swagger.DisableSwaggerSecurity;
 import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
@@ -31,6 +32,8 @@ import org.springframework.web.bind.annotation.RestController;
 public class ClubController {
 
     private final ClubService clubService;
+
+    private final ClubCategoyService clubCategoyService;
 
     @DisableSwaggerSecurity
     @Operation(summary = "분과별 중앙동아리 조회")
@@ -59,21 +62,21 @@ public class ClubController {
     @Operation(summary = "메인페이지에서의 전체 동아리 해시태그 목록 조회")
     @GetMapping("/hashtags")
     public List<EnumMapperVO> getClubsTotalHashTags() {
-        return clubService.getClubsTotalHashtags();
+        return clubCategoyService.getHashtags();
     }
 
     @DisableSwaggerSecurity
     @Operation(summary = "중앙동아리 - 분과 이름 목록 조회 \uD83D\uDCCB")
     @GetMapping("/divisions")
     public List<EnumMapperVO> getDivisions() {
-        return clubService.getDivisionNames();
+        return clubCategoyService.getDivisions();
     }
 
     @DisableSwaggerSecurity
     @Operation(summary = "소모임 - 단과대별 학과 이름 목록 조회")
     @GetMapping("/colleges")
     public List<CollegeResponse> getColleges() {
-        return clubService.getCollegesWithDepartments();
+        return clubCategoyService.getCollegesWithDepartments();
     }
 
     @DisableSwaggerSecurity
@@ -101,21 +104,21 @@ public class ClubController {
     @Operation(summary = "clubType 목록 반환 \uD83D\uDCCB")
     @GetMapping(value = "/types")
     public List<EnumMapperVO> getClubTypes() {
-        return clubService.getClubTypes();
+        return clubCategoyService.getClubTypes();
     }
 
     @DisableSwaggerSecurity
     @Operation(summary = "소모임 - 단과대 목록 조회 \uD83D\uDCCB")
     @GetMapping(value = "/colleges/types")
     public List<EnumMapperVO> getCollegeList() {
-        return clubService.getColleges();
+        return clubCategoyService.getColleges();
     }
 
     @DisableSwaggerSecurity
     @Operation(summary = "소모임 - 단과대 소속 학과 목록 조회 \uD83D\uDCCB")
     @GetMapping(value = "/departments", params = "college")
     public List<EnumMapperVO> getDepartmentList(@RequestParam(required = false) College college) {
-        return clubService.getDepartmentList(college);
+        return clubCategoyService.getDepartmentsByCollege(college);
     }
 
     /**TODO
