+++ src/main/java/com/clubber/ClubberServer/domain/club/service/ClubCategoyService.java
@@ -1,11 +1,13 @@
 package com.clubber.ClubberServer.domain.club.service;
 
 import com.clubber.ClubberServer.domain.club.domain.College;
+import com.clubber.ClubberServer.domain.club.dto.CollegeResponse;
 import com.clubber.ClubberServer.global.mapper.enums.EnumMapper;
 import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
 
 @Service
@@ -13,22 +15,32 @@ import java.util.List;
 public class ClubCategoyService {
     private final EnumMapper enumMapper;
 
-    // [회원가입] 동아리 type 목록 조회
-    public List<EnumMapperVO> getClubTypes() {
-        return enumMapper.get("ClubType");
-    }
-
     // [중앙 동아리] - 분과명 반환 (enum)
     public List<EnumMapperVO> getDivisions() {
         return enumMapper.get("Division");
     }
 
-    // [소모임] 단과대별 소속 학과 목록 조회
+    // [소모임] - 전체 단과대별 소속 학과 목록 조회
+    public List<CollegeResponse> getCollegesWithDepartments() {
+        return Arrays.stream(College.values())
+                .map(
+                        college -> {
+                            List<EnumMapperVO> enumMapperVOs = enumMapper.toEnumValues(college.getDepartments());
+                            return CollegeResponse.from(college, enumMapperVOs);
+                        }).toList();
+    }
+
+    // [회원가입] 단일 단과대 소속 학과 목록 조회
     public List<EnumMapperVO> getDepartmentsByCollege(College college) {
         return enumMapper.toEnumValues(college.getDepartments());
     }
 
-    // [회원가입] 소모임 단과대 목록 조회
+    // [회원가입] 동아리 type 목록 조회
+    public List<EnumMapperVO> getClubTypes() {
+        return enumMapper.get("ClubType");
+    }
+
+    // [회원가입] 단과대 목록 반환
     public List<EnumMapperVO> getColleges() {
         return enumMapper.get("College");
     }
